Split runtime function execution
- Move argument binding, return-body loops, and step execution into focused runtime::function modules.
- Keep runtime::function as the function-call runner facade.
- Colocate unit tests with the runtime module that owns each execution path.
- Preserve runtime function coverage and review-policy boundaries.
